+++ test/grax/rdf/preloader_test.exs
@@ -1,8 +1,8 @@
 defmodule Grax.RDF.PreloaderTest do
   use Grax.TestCase
 
-  alias Grax.RDF.Preloader
-  alias Grax.InvalidResourceTypeError
+  alias Grax.RDF.{Preloader, Loader}
+  alias Grax.{Link, InvalidResourceTypeError}
 
   test "link to itself without circle" do
     assert RDF.graph([
@@ -13,7 +13,7 @@ defmodule Grax.RDF.PreloaderTest do
            |> Example.SelfLinked.load(EX.A) ==
              Example.SelfLinked.build(EX.A,
                name: "a",
-               next: Example.SelfLinked.build!(EX.B, name: "b")
+               next: Example.SelfLinked.build!(EX.B, name: "b", next: Link.NotLoaded)
              )
   end
 
@@ -26,7 +26,7 @@ defmodule Grax.RDF.PreloaderTest do
            |> Example.SelfLinked.load(EX.A) ==
              Example.SelfLinked.build(EX.A,
                name: "a",
-               next: Example.SelfLinked.build!(~B"b", name: "b")
+               next: Example.SelfLinked.build!(~B"b", name: "b", next: Link.NotLoaded)
              )
 
     assert RDF.graph([
@@ -40,7 +40,7 @@ defmodule Grax.RDF.PreloaderTest do
                next:
                  Example.SelfLinked.build!(~B"b",
                    name: "b",
-                   next: Example.SelfLinked.build!(~B"c", name: "c", next: nil)
+                   next: Example.SelfLinked.build!(~B"c", name: "c")
                  )
              )
   end
@@ -48,7 +48,7 @@ defmodule Grax.RDF.PreloaderTest do
   test "direct link to itself" do
     assert RDF.graph([EX.A |> EX.name("a") |> EX.next(EX.A)])
            |> Example.SelfLinked.load(EX.A) ==
-             Example.SelfLinked.build(EX.A, name: "a")
+             Example.SelfLinked.build(EX.A, name: "a", next: Link.NotLoaded)
   end
 
   test "link to itself with circle" do
@@ -65,7 +65,9 @@ defmodule Grax.RDF.PreloaderTest do
                  Example.Circle.build!(EX.B,
                    name: "b",
                    link2: [],
-                   link1: [Example.Circle.build!(EX.C, name: "c", link2: [])]
+                   link1: [
+                     Example.Circle.build!(EX.C, name: "c", link1: Grax.Link.NotLoaded)
+                   ]
                  )
                ]
              )
@@ -95,7 +97,7 @@ defmodule Grax.RDF.PreloaderTest do
                          Example.Circle.build!(EX.C,
                            name: "c",
                            link2: [],
-                           link1: [Example.Circle.build!(EX.E, name: "e", link2: [])]
+                           link1: [Example.Circle.build!(EX.E, name: "e", link1: Link.NotLoaded)]
                          )
                        ]
                      )
@@ -112,7 +114,7 @@ defmodule Grax.RDF.PreloaderTest do
                          Example.Circle.build!(EX.B,
                            name: "b",
                            link2: [],
-                           link1: [Example.Circle.build!(EX.D, name: "d", link2: [])]
+                           link1: [Example.Circle.build!(EX.D, name: "d", link1: Link.NotLoaded)]
                          )
                        ]
                      )
@@ -146,7 +148,7 @@ defmodule Grax.RDF.PreloaderTest do
                          Example.Circle.build!(EX.C,
                            name: "c",
                            link2: [],
-                           link1: [Example.Circle.build!(EX.E, name: "e", link1: [])]
+                           link1: [Example.Circle.build!(EX.E, name: "e", link2: Link.NotLoaded)]
                          )
                        ]
                      )
@@ -163,7 +165,7 @@ defmodule Grax.RDF.PreloaderTest do
                          Example.Circle.build!(EX.B,
                            name: "b",
                            link2: [],
-                           link1: [Example.Circle.build!(EX.D, name: "d", link1: [])]
+                           link1: [Example.Circle.build!(EX.D, name: "d", link2: Link.NotLoaded)]
                          )
                        ]
                      )
@@ -181,8 +183,10 @@ defmodule Grax.RDF.PreloaderTest do
              ])
              |> Example.HeterogeneousLinks.load(EX.A) ==
                Example.HeterogeneousLinks.build(EX.A,
-                 one: Example.Post.build!(EX.Post1, title: "foo"),
-                 strict_one: Example.Post.build!(EX.Post1, title: "foo"),
+                 one:
+                   Example.Post.build!(EX.Post1, title: "foo") |> Loader.init_link_properties(),
+                 strict_one:
+                   Example.Post.build!(EX.Post1, title: "foo") |> Loader.init_link_properties(),
                  many: []
                )
 
@@ -192,8 +196,12 @@ defmodule Grax.RDF.PreloaderTest do
              ])
              |> Example.HeterogeneousLinks.load(EX.A) ==
                Example.HeterogeneousLinks.build(EX.A,
-                 one: Example.Comment.build!(EX.Comment1, content: "foo"),
-                 strict_one: Example.Comment.build!(EX.Comment1, content: "foo"),
+                 one:
+                   Example.Comment.build!(EX.Comment1, content: "foo")
+                   |> Loader.init_link_properties(),
+                 strict_one:
+                   Example.Comment.build!(EX.Comment1, content: "foo")
+                   |> Loader.init_link_properties(),
                  many: []
                )
 
@@ -205,7 +213,10 @@ defmodule Grax.RDF.PreloaderTest do
                Example.HeterogeneousLinks.build(EX.A,
                  one: nil,
                  strict_one: nil,
-                 many: [Example.Comment.build!(EX.Comment1, content: "foo")]
+                 many: [
+                   Example.Comment.build!(EX.Comment1, content: "foo")
+                   |> Loader.init_link_properties()
+                 ]
                )
     end
 
@@ -218,7 +229,9 @@ defmodule Grax.RDF.PreloaderTest do
                Example.HeterogeneousLinks.build(EX.A,
                  one: nil,
                  strict_one: nil,
-                 many: [Example.Post.build!(EX.Post1, title: "foo")]
+                 many: [
+                   Example.Post.build!(EX.Post1, title: "foo") |> Loader.init_link_properties()
+                 ]
                )
 
       assert RDF.graph([
@@ -231,8 +244,10 @@ defmodule Grax.RDF.PreloaderTest do
                  one: nil,
                  strict_one: nil,
                  many: [
-                   Example.Comment.build!(EX.Comment1, content: "foo"),
+                   Example.Comment.build!(EX.Comment1, content: "foo")
+                   |> Loader.init_link_properties(),
                    Example.Post.build!(EX.Post1, title: "foo")
+                   |> Loader.init_link_properties()
                  ]
                )
     end
@@ -269,7 +284,9 @@ defmodule Grax.RDF.PreloaderTest do
              ])
              |> Example.HeterogeneousLinks.load(EX.A) ==
                Example.HeterogeneousLinks.build(EX.A,
-                 one: Example.Comment.build!(EX.Comment1, content: "bar"),
+                 one:
+                   Example.Comment.build!(EX.Comment1, content: "bar")
+                   |> Loader.init_link_properties(),
                  strict_one: nil,
                  many: []
                )
@@ -320,7 +337,7 @@ defmodule Grax.RDF.PreloaderTest do
              Example.DepthPreloading.build(EX.A,
                next:
                  Example.DepthPreloading.build!(EX.B,
-                   next: Example.DepthPreloading.build!(EX.C)
+                   next: Example.DepthPreloading.build!(EX.C, next: Link.NotLoaded)
                  )
              )
   end
@@ -360,6 +377,7 @@ defmodule Grax.RDF.PreloaderTest do
                            link2: [],
                            link1: [
                              Example.Circle.build!(EX.B, name: "b")
+                             |> Loader.init_link_properties()
                            ]
                          )
                        ]
@@ -396,7 +414,10 @@ defmodule Grax.RDF.PreloaderTest do
                              Example.Circle.build!(EX.E,
                                name: "e",
                                link2: [],
-                               link1: [Example.Circle.build!(EX.B, name: "b")]
+                               link1: [
+                                 Example.Circle.build!(EX.B, name: "b")
+                                 |> Loader.init_link_properties()
+                               ]
                              )
                            ]
                          )
@@ -419,7 +440,10 @@ defmodule Grax.RDF.PreloaderTest do
                              Example.Circle.build!(EX.D,
                                name: "d",
                                link2: [],
-                               link1: [Example.Circle.build!(EX.C, name: "c")]
+                               link1: [
+                                 Example.Circle.build!(EX.C, name: "c")
+                                 |> Loader.init_link_properties()
+                               ]
                              )
                            ]
                          )